moved table schema and TestName struct from src/types.rs to src/api/types.rs. moved establish_connection() function from src/lib.rs to src/api/mod.rs
