+++ src/api/types.rs
@@ -1,0 +1,14 @@
+
+
+#[derive(Queryable, Serialize, Deserialize)]
+pub struct TestName {
+    pub id: i32,
+    pub name: String,
+}
+
+table! {
+	test_name (id) {
+		id -> Int4,
+		name -> Varchar,
+	}
+}
